Restructured Tests
Renamed and moved UnitTests to IntegrationTests
Renamed and moved DOLDatabaseTests to IntegrationTests_Database
Changed names in appveyor, travis and csproj accordingly
