+++ IntegrationTests/.gitignore
@@ -1,0 +1,3 @@
+/bin
+/obj
+/IntegrationTests.xml
