+++ IntegrationTests/Managers/FriendsManagerTest.cs
