+++ IntegrationTests/Managers/InvalidNamesManagerTest.cs
