+++ IntegrationTests/Managers/WeatherManagerTest.cs
