+++ IntegrationTests_Database/Properties/AssemblyInfo.cs
