+++ IntegrationTests_Database/RegisterTableTests.cs
