--- UnitTests/.gitignore
@@ -1,3 +1,0 @@
-/bin
-/obj
-/UnitTests.xml
